# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateTransactionRequest:
    properties:
      cinema_id:
        type: integer
      date_booking:
        type: string
      location:
        type: string
      movie_id:
        type: integer
      payment_methods_id:
        type: integer
      seats:
        items:
          type: string
        type: array
      status:
        type: string
      time_booking:
        type: string
      total_price:
        type: integer
      user_id:
        type: string
    type: object
...
  dto.UpdateProfileRequest:
    properties:
      firstname:
        type: string
      gender:
        type: string
      lastname:
        type: string
      phone_number:
        type: string
      profile_picture:
        type: string
      username:
        type: string
    type: object
  utils.Response:
    properties:
      errors:
        type: string
      message:
        type: string
      otp:
        type: string
      result: {}
      status:
        type: integer
      success:
        type: boolean
      token:
        type: string
      total:
        type: integer
    type: object
info:
  contact: {}
  description: This is a backend service for tickitz web app
  title: Cinemax API
  version: "1.0"
paths:
...
  /profile:
    get:
      consumes:
      - application/json
      description: Get profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Token: []
      summary: Handle get profile
      tags:
      - profile
    patch:
      consumes:
      - application/json
      description: Update profile
      parameters:
      - description: request update profile
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Token: []
      summary: Handle update profile
      tags:
      - profile
    post:
      consumes:
      - multipart/form-data
      description: Upload photo
      parameters:
      - description: photo
        in: formData
        name: photo
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Token: []
      summary: Handle upload photo
      tags:
      - profile
  /transactions:
    post:
      consumes:
      - application/json
      description: Create a new transaction
      parameters:
      - description: request create transaction
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Handle create transaction
      tags:
      - transactions
  /transactions/history:
    get:
      consumes:
      - application/json
      description: Get history
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Token: []
      summary: Handle get history
      tags:
      - profile
securityDefinitions:
  Token:
    description: |-
      Use Bearer token to access protected routes
      Example: Bearer your_token_here
      Make sure to include the word "Bearer" followed by a space before the token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"